+++ test/module/logger.test.ts
@@ -11,7 +11,7 @@ describe("KipperLogger", () => {
 
       // Simple message logging
       logger.info("A message");
-    })
+    });
 
     it("Construction with a basic emit handler", () => {
       let called: number = 0;
@@ -28,12 +28,12 @@ describe("KipperLogger", () => {
 
       logger.info("Another message");
       assert(called === 2, "Function must have been called twice");
-    })
+    });
   });
 
   describe("KipperLogger.log()", () => {
     it("Simple call", () => {
-      let levelReceived: LogLevel = 0;
+      let levelReceived: LogLevel = LogLevel.UNKNOWN;
       const emitHandler = (level: LogLevel) => {
         levelReceived = level;
       };
@@ -43,22 +43,28 @@ describe("KipperLogger", () => {
 
       // Simple message logging
       logger.debug("A message");
+      // @ts-ignore
       assert(levelReceived === LogLevel.DEBUG, "Level must be the same");
 
       logger.info("A message");
+      // @ts-ignore
       assert(levelReceived === LogLevel.INFO, "Level must be the same");
 
       logger.warn("A message");
+      // @ts-ignore
       assert(levelReceived === LogLevel.WARN, "Level must be the same");
 
       logger.error("A message");
+      // @ts-ignore
       assert(levelReceived === LogLevel.ERROR, "Level must be the same");
 
       logger.fatal("A message");
+      // @ts-ignore
       assert(levelReceived === LogLevel.FATAL, "Level must be the same");
 
       logger.trace("A message");
+      // @ts-ignore
       assert(levelReceived === LogLevel.TRACE, "Level must be the same");
-    })
-  })
+    });
+  });
 });